docs: fix examples
