+++ README.md
@@ -6,11 +6,10 @@ Go SDK to enable your Go application to create GitHub User Access Tokens for Git
 
 ## Examples
 
-- [Using configuration file and keyring](examples/simple-2/main.go)
-- [Passing a client id without configuration file and keyring](examples/simple-1/main.go)
-- [Customizing Logging](examples/simple-5/main.go)
-- [Customizing opening the browser](examples/simple-4/main.go)
-- [Customizing showing the device code](examples/simple-4/main.go)
+- [Using configuration file and keyring](examples/simple-1/main.go)
+- [Customizing Logging](examples/simple-4/main.go)
+- [Customizing opening the browser](examples/simple-3/main.go)
+- [Customizing showing the device code](examples/simple-3/main.go)
 
 ## Using logging libraries such as logrus, zap, and zerolog
